Issue #81: Split test classes
