+++ tests/functional/SingleListingPageTest.php
@@ -1,149 +1,10 @@
 <?php
 
+
 namespace MarketBoard\Tests\Functional;
 
-class ListingPageTest extends BaseTestCase
+class SingleListingPageTest extends BaseTestCase
 {
-    private static $listing_data = [
-        [
-            "email" => "test@test.com",
-            "subcategory_id" => null,
-            "unit_price" => "123",
-            "quantity" => "2",
-            "removal_code" => "AAAAAA",
-            "description" => "Lorem Ipsum1",
-            "created_at" => "2020-06-18 23:14:18",
-        ],
-        [
-            "email" => "test3@test.com",
-            "subcategory_id" => null,
-            "unit_price" => "789",
-            "quantity" => "1",
-            "removal_code" => "AAAAAA",
-            "description" => "Lorem Ipsum2",
-            "created_at" => "2020-06-18 23:14:17",
-        ],
-        [
-            "email" => "test2@test.com",
-            "subcategory_id" => null,
-            "unit_price" => "456",
-            "quantity" => "3",
-            "removal_code" => "AAAAAA",
-            "description" => "Lorem Ipsum3",
-            "created_at" => "2020-06-18 23:14:16",
-        ]
-    ];
-
-    private static $category = [
-        "category_name" => "category_test"
-    ];
-
-    private static $subcategory = [
-        "subcategory_name" => "subcategory_test",
-        "category_id" => null
-    ];
-
-    public static function setUpBeforeClass(): void
-    {
-        $query = "INSERT INTO categories(category_name) VALUES(?);";
-        $statement1 = self::$container['db']->prepare($query);
-        $statement1->execute(array_values(self::$category));
-        $categoryId = self::$container['db']->lastInsertId();
-        self::$category["id"] = $categoryId; // Set id so clearDatabaseOf() only removes one entry
-
-        self::$subcategory["category_id"] = $categoryId;
-        $query = "INSERT INTO subcategories(subcategory_name, category_id) VALUES(?, ?);";
-        $statement2 = self::$container['db']->prepare($query);
-        $statement2->execute(array_values(self::$subcategory));
-        $subcategoryId = self::$container['db']->lastInsertId();
-
-        self::$listing_data[0]["subcategory_id"] = $subcategoryId;
-        self::$listing_data[1]["subcategory_id"] = $subcategoryId;
-    }
-
-    public function setUp(): void
-    {
-        $this->clearLog();
-    }
-
-    public function tearDown(): void
-    {
-        $this->clearLog();
-    }
-
-    public static function tearDownAfterClass(): void
-    {
-        self::clearListingsTable();
-
-        unset(self::$subcategory['category_id']);
-        self::clearDatabaseOf("subcategories", self::$subcategory);
-        unset(self::$category['id']);
-        self::clearDatabaseOf("categories", self::$category);
-        self::$category['id'] = null;
-    }
-
-    /**
-     * Verify correct html element are displayed when
-     * viewing webpage on  '/listings/new'
-     */
-    public function testGETNewListing()
-    {
-        $response = $this->processRequest('GET', '/listings/new');
-        $this->assertEquals(200, $response->getStatusCode());
-
-        $this->assertLogDoesNotContain(['ERROR']);
-
-        $htmlBody = (string)$response->getBody();
-
-        // Verify input fields
-        $this->assertStringContainsString('<input type="email" class="form-control" id="new_listing_email" value="" placeholder="Enter your E-mail" name="email" required>', $htmlBody);
-    }
-
-    /**
-     * Verify that values are correctly inserted into the database
-     * when received as POST on '/listings/new'
-     */
-    public function testPOSTNewListing()
-    {
-        $listing_data = self::$listing_data[0];
-        unset($listing_data['removal_code']);
-        unset($listing_data['created_at']);
-        $listing_data['captcha'] = self::$container['session']->get('captcha');
-        $response = $this->processRequest('POST', '/listings/new', $listing_data);
-        unset($listing_data['captcha']);
-        $this->assertEquals(200, $response->getStatusCode());
-
-        $this->assertLogDoesNotContain(['ERROR']);
-        $this->assertLogContains(["INFO: Parameters inserted"]);
-
-        $this->verifyEntryInserted("listings", $listing_data);
-    }
-
-    /**
-     * Verify that multiple listings are displayed when
-     * viewing webpage on  '/listings/listings/'
-     */
-    public function testGETAllListings()
-    {
-        $query = "INSERT INTO listings(email, subcategory_id, unit_price, quantity, removal_code, description, created_at) VALUES(?,?,?,?,?,?,?);";
-        $statement1 = self::$container['db']->prepare($query);
-        $statement1->execute(array_values(self::$listing_data[0]));
-        $statement2 = self::$container['db']->prepare($query);
-        $statement2->execute(array_values(self::$listing_data[1]));
-
-        $response = $this->processRequest('GET', '/listings/');
-        $this->assertEquals(200, $response->getStatusCode());
-
-        $this->assertLogDoesNotContain(['ERROR']);
-
-        $htmlBody = (string)$response->getBody();
-
-        // Verify input fields
-        $this->assertStringContainsString(self::$listing_data[0]['unit_price'], $htmlBody);
-        $this->assertStringContainsString(self::$listing_data[1]['unit_price'], $htmlBody);
-        $this->assertStringContainsString(self::$subcategory['subcategory_name'], $htmlBody);
-    }
-
     /**
      * Verify that a single listing are displayed when
      * viewing webpage on  '/listings/{id}/'
